Cleaned up tests
Created a QuestionnaireLogicTests class to test the QUestionnaireLogicTest

Removed other two files
